requirements.txt
# conda create -n no_pain_diffusion python=3.10 -y
# conda activate no_pain_diffusion

# torch (建议先单独安装)
# torch==2.3.1
# torchvision==0.18.1

# torch util
einops==0.8.0
# deepspeed==0.8.3  # 分布式训练优化

# timm
# timm==0.6.13
timm==1.0.19

# diffusers
huggingface-hub==0.25.1
accelerate==1.2.1
diffusers==0.32.1
transformers==4.47.1
tokenizers==0.21.0
sentencepiece==0.2.0
protobuf==3.20.3
safetensors==0.4.5
datasets==3.4.1

# metric
opencv-python==4.11.0.86
# opencv-contrib-python==4.11.0.86
# hpsv2==1.2.0
# pytorch_fid==0.3.0
# pytorch_image_generation_metrics==0.6.1
scikit-image==0.25.0
# lpips==0.1.4
# open-clip-torch==2.23.0
# image-reward==1.5
# git+https://github.com/openai/CLIP.git

# Hydra
hydra-core==1.3.2
omegaconf==2.3.0
# pickle5==0.0.11  # Python 自带 `pickle`, 无需安装
PyYAML==6.0.2

# misc
numpy==2.1.3
matplotlib==3.9.2
pandas==2.2.3
Pillow==10.1.0
tqdm==4.67.1
requests==2.32.3
# rich==13.3.2  # 终端美化
